fix: 자세 경고 반복 미작동 버그 수정 feat: 자세 경고 시연용 임계값 단축 (5/30/180s → 5/10/15s), PiP 화면 경고 단계별 메시지 표시 추가, WorkCam 자세 상태 문구 개선

+++ backend/app/models/posture_log.py
@@ -1,6 +1,9 @@
 from sqlalchemy import Column, Integer, Float, Boolean, DateTime, ForeignKey, JSON
 from datetime import datetime
 from app.database import Base
+from sqlalchemy import BigInteger
+
+
 
 class PostureLog(Base):
     __tablename__ = "posture_logs"
@@ -10,7 +13,7 @@ class PostureLog(Base):
     session_id = Column(Integer, ForeignKey("sessions.session_id"), nullable = False)
     calibration_id = Column(Integer, ForeignKey("calibrations.calibration_id"), nullable = False)
 
-    timestamp_ms = Column(Integer, nullable = False)
+    timestamp_ms = Column(BigInteger, nullable=False)
 
     features = Column(JSON, nullable = False)
